requirements.txt
fastapi
uvicorn
google-generativeai 
psycopg2-binary 
google-auth 
google-auth-oauthlib 
google-api-python-client 
python-dotenv